requirements.txt
asgiref==3.11.1
attrs==26.1.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
Django==4.2.11
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
feedparser==6.0.12
google-genai==1.68.0
idna==3.11
jsonschema==4.26.0
openai==2.30.0
python-dotenv==1.2.2
requests==2.33.0
sendgrid==6.12.5
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
